[Fix #146109503] install sequelize-cli to fix failing travis
