+++ package.json
@@ -62,6 +62,7 @@
     "nyc": "^10.3.0",
     "sass-loader": "^3.1.2",
     "script-loader": "^0.6.1",
+    "sequelize-cli": "^2.7.0",
     "style-loader": "^0.13.0",
     "supertest": "^3.0.0",
     "webpack": "^1.12.13"
